test(backend): correctly run all it tests with "test:it" script
