+++ backend/package.json
@@ -31,7 +31,7 @@
     "dev": "nest start --watch",
     "build": "nest build",
     "start": "node dist/main.js",
-    "test": "jest --config ./jest.config.ts",
+    "test:it": "jest --config ./jest.config.ts",
     "test:e2e": "cross-env NODE_ENV=test NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" jest --config ./test/jest-e2e.config.ts",
     "db:reset": "dotenv -e ../.env -- sh -c 'pnpm prisma migrate reset && pnpm db:seed:dev'",
     "db:push": "dotenv -e ../.env -- pnpm prisma db push",
